+++ pypot/dynamixel/__init__.py
@@ -1,12 +1,19 @@
+import sys
 import platform
 import glob
+import logging
+
+import serial.tools.list_ports
+
 
 from .io import DxlIO, Dxl320IO, DxlError
 from .error import BaseErrorHandler
-from .controller import BaseDxlController
+from .syncloop import BaseDxlController
 from .motor import DxlMXMotor, DxlAXRXMotor, DxlXL320Motor
 
 from ..robot import Robot
+
+logger = logging.getLogger(__name__)
 
 
 def _get_available_ports():
@@ -16,6 +23,9 @@ def _get_available_ports():
 
     elif platform.system() == 'Linux':
         return glob.glob('/dev/ttyACM*') + glob.glob('/dev/ttyUSB*')
+
+    elif sys.platform.lower() == 'cygwin':
+        return glob.glob('/dev/com*')
 
     elif platform.system() == 'Windows':
         import _winreg
@@ -30,7 +40,8 @@ def _get_available_ports():
                 ports.append(str(_winreg.EnumValue(key, i)[1]))
             except WindowsError:
                 return ports
-
+    else:
+        raise EnvironmentError('{} is an unsupported platform, cannot find serial ports !'.format(platform.system()))
     return []
 
 
@@ -43,6 +54,22 @@ def get_available_ports(only_free=False):
     return ports
 
 
+def get_port_vendor_info(port=None):
+    """ Return vendor informations of a usb2serial device.
+        It may depends on the Operating System.
+        :param string port: port of the usb2serial device
+
+        :Example:
+
+        Result with a USB2Dynamixel on Linux:
+        In [1]: import pypot.dynamixel
+        In [2]: pypot.dynamixel.get_port_vendor_info('/dev/ttyUSB0')
+        Out[2]: 'USB VID:PID=0403:6001 SNR=A7005LKE' """
+
+    port_info_dict = dict((x[0], x[2]) for x in serial.tools.list_ports.comports())
+    return port_info_dict[port] if port is not None else port_info_dict
+
+
 def find_port(ids, strict=True):
     """ Find the port with the specified attached motor ids.
 
@@ -52,21 +79,30 @@ def find_port(ids, strict=True):
         .. warning:: If two (or more) ports are attached to the same list of motor ids the first match will be returned.
 
     """
+    ids_founds = []
     for port in get_available_ports():
         for DxlIOCls in (DxlIO, Dxl320IO):
             try:
                 with DxlIOCls(port) as dxl:
-                    founds = len(dxl.scan(ids))
+                    _ids_founds = dxl.scan(ids)
+                    ids_founds += _ids_founds
 
-                    if strict and founds == len(ids):
+                    if strict and len(_ids_founds) == len(ids):
                         return port
 
-                    if not strict and founds >= len(ids) / 2:
+                    if not strict and len(_ids_founds) >= len(ids) / 2:
+                        logger.warning('Missing ids: {}'.format(ids, list(set(ids) - set(_ids_founds))))
                         return port
+
+                    if len(ids_founds) > 0:
+                        logger.warning('Port:{} ids found:{}'.format(port, _ids_founds))
+
             except DxlError:
+                logger.warning('DxlError on port {}'.format(port))
                 continue
 
-    raise IndexError('No suitable port found for ids {}!'.format(ids))
+    raise IndexError('No suitable port found for ids {}. These ids are missing {} !'.format(
+        ids, list(set(ids) - set(ids_founds))))
 
 
 def autodetect_robot():