+++ pypot/primitive/utils.py
@@ -3,13 +3,11 @@ import numpy
 from copy import deepcopy
 from collections import defaultdict
 
-from .primitive import LoopPrimitive
+from .primitive import Primitive, LoopPrimitive
 
 
 class Sinus(LoopPrimitive):
-
-    """ Apply a sinus on the motor specified as argument. Parameters (amp, offset and phase) should be specified in degree
-    """
+    """ Apply a sinus on the motor specified as argument. Parameters (amp, offset and phase) should be specified in degree. """
     properties = LoopPrimitive.properties + ['frequency', 'amplitude', 'offset', 'phase']
 
     def __init__(self, robot, refresh_freq,
@@ -67,9 +65,7 @@ class Sinus(LoopPrimitive):
 
 
 class Cosinus(Sinus):
-
-    """ Apply a cosinus on the motor specified as argument. Parameters (amp, offset and phase) should be specified in degree
-    """
+    """ Apply a cosinus on the motor specified as argument. Parameters (amp, offset and phase) should be specified in degree. """
 
     def __init__(self, robot, refresh_freq,
                  motor_list,
@@ -122,6 +118,7 @@ class PositionWatcher(LoopPrimitive):
 
         self._pos = defaultdict(list)
         self.watched_motors = watched_motors
+        self._duration = 0.
 
     @property
     def record_positions(self):
@@ -134,7 +131,35 @@ class PositionWatcher(LoopPrimitive):
         for m in self.watched_motors:
             self._pos[m.name].append(m.present_position)
 
+        self._duration = self.elapsed_time
+
     def plot(self, ax):
         for m, pos in self._pos.items():
-            ax.plot(pos)
-        ax.legend(self._pos.keys())
+            t = numpy.linspace(0, self._duration, len(pos))
+            ax.plot(t, pos)
+        ax.set_ylabel('position (in deg)')
+        ax.set_xlabel('time (in s)')
+        ax.legend(self._pos.keys(), loc='best')
+
+
+class SimplePosture(Primitive):
+    def __init__(self, robot, duration):
+        Primitive.__init__(self, robot)
+
+        self.duration = duration
+
+    def setup(self):
+        self._speeds = {m: m.moving_speed for m in self.robot.motors}
+
+    def run(self):
+        if not hasattr(self, 'target_position'):
+            raise NotImplementedError('You have to define "target_position" first!')
+
+        for m in self.robot.motors:
+            m.compliant = False
+
+        self.robot.goto_position(self.target_position, self.duration, wait=True)
+
+    def teardown(self):
+        for m, s in self._speeds.items():
+            m.moving_speed = s