+++ pypot/sensor/__init__.py
@@ -1,1 +1,2 @@
 from camera import *
+from imagefeature import *